add force refresh program to API, workaround - clear DB
